email_assistant/pyproject.toml
[project]
name = "emailAssistant"
version = "0.1.0"
description = "Assistant sends PDF to Customers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [    
    "ollama>=0.4.5",
    "openai>=1.0.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "rank-bm25>=0.2.2",
    "uvicorn>=0.32.0",
    "msal>=1.30.0",
    "requests>=2.32.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "pypdf>=4.3.1",
    "pdfplumber>=0.11.0",
]
